At least one user will survive from now on
Client code checks to make sure current user cannot delete itself, and
server side code makes sure there is at least 1 user left after this
delete happens. Else the delete request will not be processed.

+++ src/edu/cmu/sv/fsgim/business/service/UserManager.java
@@ -104,7 +104,15 @@ public class UserManager {
 					e);
 		}
 
-		boolean isDeleted = dao.delete(userId);
+		boolean isDeleted = false;
+		synchronized("UserManager") {
+			List<UserPO> users = dao.findAll();
+			if(users == null || users.size() <= 1) {
+				throw new RuntimeException("Trying to delete the last "
+						+ "available user is not supported.");
+			}
+			isDeleted = dao.delete(userId);	
+		}
 
 		// If there is something wrong, and the query did not get deleted,
 		// raise an exception and report it.
